+++ frt_datapackage/scripts/FRT_Stan_in_Parallel.R
@@ -1,7 +1,7 @@
 ###########################
 # Run model in parallel
 # Christopher Gandrud
-# 16 February 2014
+# 27 March 2015
 # MIT License
 ###########################
 
@@ -79,10 +79,8 @@ sflist <-
     mclapply(1:4, mc.cores = 4,
         function(i) stan(fit = empty_stan, data = frt_data,
                         seed = i, chains = 1,
-                        iter = 10000, chain_id = i,
-                        pars = c('delta', 'alpha', 'beta', 'log_gamma'),
-                        diagnostic_file = paste0(
-                            'frt_sims_diagnostic', Sys.Date())
+                        iter = 20000, chain_id = i,
+                        pars = c('delta', 'alpha', 'beta', 'log_gamma')
         )
     )